Added template for websocket tests
