+++ src/test/java/com/cgm/infolab/WebSocketDeleteEditPublicTests.java
@@ -1,28 +1,19 @@
 package com.cgm.infolab;
 
-import com.cgm.infolab.configuration.TestSecurityConfiguration;
 import com.cgm.infolab.db.model.ChatMessageEntity;
 import com.cgm.infolab.db.model.RoomEntity;
 import com.cgm.infolab.db.model.UserEntity;
 import com.cgm.infolab.db.model.Username;
-import com.cgm.infolab.helper.TestDbHelper;
-import com.cgm.infolab.helper.TestStompHelper;
 import com.cgm.infolab.model.ChatMessageDto;
 import com.cgm.infolab.model.WebSocketMessageDto;
+import com.cgm.infolab.templates.WebSocketTest;
 import org.apache.commons.lang3.tuple.Pair;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestInstance;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.messaging.simp.stomp.StompFrameHandler;
 import org.springframework.messaging.simp.stomp.StompHeaders;
 import org.springframework.messaging.simp.stomp.StompSession;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.web.socket.messaging.WebSocketStompClient;
 
 import java.lang.reflect.Type;
 import java.time.LocalDateTime;
@@ -35,33 +26,16 @@ import static com.cgm.infolab.model.WebSocketMessageTypeEnum.DELETE;
 import static com.cgm.infolab.model.WebSocketMessageTypeEnum.EDIT;
 import static org.awaitility.Awaitility.await;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@ActiveProfiles({ProfilesConstants.TEST})
-@TestInstance(TestInstance.Lifecycle.PER_CLASS)
-public class WebSocketDeleteEditPublicTests {
-    @LocalServerPort
-    public Integer port;
-    @Autowired
-    public TestDbHelper testDbHelper;
-    @Autowired
-    public TestStompHelper testStompHelper;
-
-    WebSocketStompClient websocket;
-
+public class WebSocketDeleteEditPublicTests extends WebSocketTest {
     UserEntity userBanana = UserEntity.of(Username.of("banana"));
-    UserEntity user1 = UserEntity.of(Username.of("user1"));
-
     RoomEntity general = RoomEntity.general();
 
-    // This is here because the @Import annotation does not work
-    @TestConfiguration
-    public static class SecurityConfiguration extends TestSecurityConfiguration {}
-
+    @Override
     @BeforeAll
-    public void setupAll(){
-        testDbHelper.clearDbExceptForGeneral();
+    public void setUp(){
+        super.setUp();
 
-        websocket = testStompHelper.initWebsocket();
+        testDbHelper.addRooms(RoomEntity.general());
 
         testDbHelper.addUsers(user1, userBanana);
